# neon base64: vectorize with vector factor 16