@sebpop 2018-02-20
Here are the numbers for the chromium_base64 implementation on the same A72 processor:

-----------------------------------------------------------------------------
Benchmark                                      Time           CPU Iterations
-----------------------------------------------------------------------------
chromium_base64_decode_lena               268360 ns     268355 ns       2608
chromium_base64_decode_peppers             23784 ns      23784 ns      29432
chromium_base64_decode_mandril            623872 ns     623852 ns       1121
chromium_base64_decode_moby_dick            2814 ns       2814 ns     248726
chromium_base64_decode_googlelogo           5936 ns       5935 ns     117943
chromium_base64_decode_bingsocialicon       3426 ns       3426 ns     204350
chromium_base64_decode_all                930364 ns     930343 ns        751