@lemire 2018-02-20
So this patch is a negative? If so, any chance you might *add* it (with new function names) instead of replacing the code we have in place? In this way, other people could inspect your work and improve upon it.