@sebpop 2018-02-20
VF16 should be a win over VF8.  I will fix LLVM to produce byte vzip and vuzp instead of the 16 byte moves.
As you suggest I will send a patch that keeps both VF8 and VF16 in different functions.