@sebpop 2018-03-01
Daniel, do you still want the two functions vf8 and vf16 side by side?
The amended patch for vf16 is now better than vf8 on all aarch64 cpus I have access to:
see commit message for 739e5fd.