@lemire 2018-03-01
Ah. Ok. So you are recommending an as-is merge then, right? Just to be clear.