@sebpop 2018-03-01
Yes, please merge in the patch as-is: the new implementation is better across the board now.

On the LLVM side: I have submitted one of the two patches needed to get the original slower patch faster https://reviews.llvm.org/D43903 the other change will be a bit more complicated as it disables a feature in instruction selection that was added for x86_64 blend instructions.  I will have to make that feature specific for x86 and disabled for aarch64.  The good thing is that the current shape of the code for vf16 does not expose the slow patterns anymore and we don't need the changes to LLVM to achieve a reasonable speedup.