@sebpop 2018-03-05
For reference https://reviews.llvm.org/D44118 is the second patch to get LLVM to produce a decent code for aarch64: this avoids the generation of many byte movs and replaces them with a zip1 and zip2 instructions.